airflow-docker/requirements.txt
mysql-connector-python==8.0.31
apache-airflow-providers-amazon==6.1.0